Compilation headers bug
The common ECC functions (usable on any platform)
never included `pirate.h`.  This caused conditionals to build incorrectly, such that `RPI_PLATFORM` was not defined.   Unfortunately, the line `#if RPI_PLATFORM == 2040` was resolving as true ... and thus compilation errors on BP6 / BP5XL / etc.

TODO: Investigate and fix the conditional compilation to first verify that `RPI_PLATFORM` is defined.....